Python-Server/requirements.txt
numpy==1.24.3
pandas==2.0.3
tensorflow==2.13.0
scikit-learn==1.3.0
python-binance==1.0.19
flask==2.3.3 